+++ closepair/util.cpp
@@ -4,11 +4,11 @@
 #include "point.hpp"
 #include "util.hpp"
 
-std::vector<Point> getRandPoints(int N, int max)
+std::vector<Point> getRandPoints(int N, long double max)
 {
     std::random_device rd;
     std::mt19937 e2(rd());
-    std::uniform_real_distribution<double> dist(0, max);
+    std::uniform_real_distribution<long double> dist(0, max);
     std::vector<Point> points;
     for (int i = 0; i < N; i++) {
         points.push_back(Point(dist(e2), dist(e2)));
